requirements.txt
streamlit~=1.45.1
pillow~=11.1.0